+++ engine/oneregion.go
@@ -2,6 +2,7 @@ package engine
 
 import (
 	"fmt"
+
 	"github.com/mumax/3/v3/cuda"
 	"github.com/mumax/3/v3/data"
 	"github.com/mumax/3/v3/util"
@@ -54,7 +55,7 @@ func (q *oneReg) Slice() (*data.Slice, bool) {
 	src := ValueOf(q.parent)
 	defer cuda.Recycle(src)
 	out := cuda.Buffer(q.NComp(), q.Mesh().Size())
-	cuda.RegionSelect(out, src, regions.Gpu(), byte(q.region))
+	cuda.RegionSelect(out, src, regions.Gpu(), uint16(q.region))
 	return out, true
 }
